Use class_create only with AJA_CREATE_DEVICE_NODES
Since class_create/destroy are GPL-only symbols, they are
not permitted to be used with a 5.9+ kernel when the
proprietary nvidia_p2p symbols from a closed source NVIDIA
driver are also used for RDMA support. This change makes it
so that class_create/destroy are only called when
AJA_CREATE_DEVICE_NODES is enabled, allowing AJA_RDMA to be
enabled for 5.9+ kernels with closed source NVIDIA drivers.

Note that enabling both AJA_RDMA and AJA_CREATE_DEVICE_NODES
on 5.9+ kernels requires the use of an open source NVIDIA
module (https://github.com/NVIDIA/open-gpu-kernel-modules).